requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
neo4j==6.1.0
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.5
typing_extensions==4.15.0
